+++ mycodo/scripts/dependencies.sh
@@ -24,9 +24,8 @@ case "${1}" in
         if [ ! -e "${INSTALL_DIRECTORY}"/env/bin/python3 ]; then
             printf "\n## Error: Virtualenv doesn't exist. Creating...\n"
             /bin/bash "${INSTALL_DIRECTORY}"/mycodo/scripts/upgrade_commands.sh setup-virtualenv
-        else
-            "${INSTALL_DIRECTORY}"/env/bin/python -m pip install --upgrade "${2}"
         fi
+        "${INSTALL_DIRECTORY}"/env/bin/python -m pip install --upgrade "${2}"
     ;;
     'internal')
         case "${2}" in