+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix PID Controller startup status error
  - Fix PCF8574 Pump Output always saving output amount to channel 0 ([#1281](https://github.com/kizniche/Mycodo/issues/1281))
  - Fix numpy import error by installing libatlas-base-dev
+ - Fix dependency install for bash-commands
 
 
 ## 8.15.7 (2023-03-02)